# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.AuthRequest:
    properties:
      authorization_code:
        type: string
      state_id:
        type: string
    type: object
  domain.AuthResponse:
    properties:
      client_origin:
        type: string
      token:
        type: string
    type: object
...
  domain.CreateDayOffRequest:
    properties:
      date_from:
        type: integer
      date_to:
        type: integer
      description:
        type: string
    type: object
  domain.CreateDayOffResponse:
    properties:
      id:
        type: string
    type: object
  domain.CreateReportRequest:
    properties:
      body:
        type: string
      break_time:
        type: integer
      date:
        type: integer
      display_name:
        type: string
      end_time:
        type: integer
      start_time:
        type: integer
      work_time:
        type: integer
    type: object
  domain.CreateReportResponse:
    properties:
      id:
        type: string
    type: object
  domain.CreateSickLeaveRequest:
    properties:
      date_from:
        type: integer
      date_to:
        type: integer
      description:
        type: string
    type: object
  domain.CreateSickLeaveResponse:
    properties:
      id:
        type: string
    type: object
...
  domain.GetCalendarEventsRequest:
    properties:
      allowed_to:
        type: string
      date_from:
        type: string
      date_to:
        type: string
      event_type:
        type: string
      page:
        type: integer
      page_size:
        type: integer
    type: object
  domain.GetCalendarEventsResponse:
    properties:
      count:
        type: integer
      events:
        items:
          $ref: '#/definitions/domain.CalendarEvent'
        type: array
    type: object
  domain.GetDayOffResponse:
    properties:
      sick_leave:
        $ref: '#/definitions/domain.DayOff'
    type: object
  domain.GetProfileRequest:
    type: object
  domain.GetProfileResponse:
    properties:
      display_name:
        type: string
      id:
        type: string
      login:
        type: string
    type: object
  domain.GetProviderUriRequest:
    properties:
      client_origin:
        type: string
    type: object
  domain.GetProviderUriResponse:
    properties:
      uri:
        type: string
    type: object
  domain.GetReportResponse:
    properties:
      report:
        $ref: '#/definitions/domain.Report'
    type: object
  domain.GetReportsRequest:
    properties:
      allowed_to:
        type: string
      date_from:
        type: string
      date_to:
        type: string
      page:
        type: integer
      page_size:
        type: integer
    type: object
  domain.GetReportsResponse:
    properties:
      count:
        type: integer
      reports:
        items:
          $ref: '#/definitions/domain.ReportItem'
        type: array
    type: object
  domain.GetSickLeaveResponse:
    properties:
      sick_leave:
        $ref: '#/definitions/domain.SickLeave'
    type: object
...
  domain.SendEmailRequest:
    properties:
      body:
        type: string
      email:
        type: string
      password:
        type: string
      recipients:
        items:
          type: string
        type: array
      subject:
        type: string
    type: object
...
  domain.SignInRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  domain.SignInResponse:
    properties:
      token:
        type: string
    type: object
...
  domain.UpdateProfileRequest:
    properties:
      display_name:
        type: string
    type: object
  domain.UpdateReportRequest:
    properties:
      body:
        type: string
      break_time:
        type: integer
      date:
        type: integer
      display_name:
        type: string
      end_time:
        type: integer
      start_time:
        type: integer
      state:
        type: string
      work_time:
        type: integer
    type: object
...
  httperror.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/httperror.ErrorResponseError'
    type: object
...
info:
  contact: {}
  description: This is a report server.
  title: Reporter API
  version: 0.0.1
paths:
  /api/v1/auth:
    get:
      description: Auth
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.AuthRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Auth
      tags:
      - Auth
  /api/v1/auth/provider:
    post:
      description: Get Provider Uri
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.GetProviderUriRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetProviderUriResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Get Provider Uri
      tags:
      - Auth
  /api/v1/calendar:
    get:
      description: Get Calendar Events
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.GetCalendarEventsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetCalendarEventsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Get Calendar Events
      tags:
      - Calendar
  /api/v1/day_offs:
    post:
      description: Create day off
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateDayOffRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.CreateDayOffResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Create day off
      tags:
      - DayOffs
  /api/v1/day_offs/{day_off_id}:
    delete:
      description: Delete day off
      parameters:
      - description: day_off_id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Delete day off
      tags:
      - DayOffs
    get:
      description: get day off
      parameters:
      - description: day_off_id
        in: path
        name: day_off_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetDayOffResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Get day off
      tags:
      - DayOffs
  /api/v1/export/reports:
    get:
      description: Export reports to csv
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.GetReportsRequest'
      responses:
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Export reports to csv
      tags:
      - Reports
  /api/v1/profile:
    get:
      description: Getting user data
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.GetProfileRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetProfileResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Get Profile
      tags:
      - Profile
    put:
      description: Updating user data
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProfileRequest'
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Update Profile
      tags:
      - Profile
  /api/v1/reports:
    get:
      description: get all reports
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.GetReportsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetReportsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Get all reports
      tags:
      - Reports
    post:
      description: Create report
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateReportRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.CreateReportResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Create report
      tags:
      - Reports
  /api/v1/reports/{report_id}:
    delete:
      description: Delete report
      parameters:
      - description: report_id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Delete report
      tags:
      - Reports
    get:
      description: get report
      parameters:
      - description: report_id
        in: path
        name: report_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetReportResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Get report
      tags:
      - Reports
    put:
      description: Update report
      parameters:
      - description: report_id
        in: path
        name: id
        required: true
        type: string
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateReportRequest'
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Update report
      tags:
      - Reports
  /api/v1/send_mail:
    post:
      description: Send email
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SendEmailRequest'
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Send email
      tags:
      - Email
  /api/v1/sick_leaves:
    post:
      description: Create sick leave
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateSickLeaveRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.CreateSickLeaveResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Create sick leave
      tags:
      - SickLeaves
  /api/v1/sick_leaves/{sick_leave_id}:
    delete:
      description: Delete sick leave
      parameters:
      - description: sick_leave_id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Delete sick leave
      tags:
      - SickLeaves
    get:
      description: get sick leave
      parameters:
      - description: sick_leave_id
        in: path
        name: sick_leave_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetSickLeaveResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Get sick leave
      tags:
      - SickLeaves
  /api/v1/sign_in:
    post:
      description: Getting an authorization token
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SignInRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SignInResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Sign In
      tags:
      - Auth
  /api/v1/sign_up:
    post:
      description: User registration
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SignUpRequest'
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Sign Up
      tags:
      - Auth
  /api/v1/stats:
    get:
      description: Get Statistics
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.GetStatisticsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetStatisticsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Get Statistics
      tags:
      - Statistics
  /api/v1/users:
    get:
      description: Get Users
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.GetUsersRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetUsersResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Get Users
      tags:
      - Users
  /api/v1/users/{user_id}/permissions:
    delete:
      description: Remove the user's permission to read an object
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.AddObjectToUserPermissionRequest'
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Remove object from user permission
      tags:
      - Permission
    post:
      description: Give the user permission to read an object
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.AddObjectToUserPermissionRequest'
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Add object to user permission
      tags:
      - Permission
  /api/v1/vacations_paid:
    post:
      description: Create vacation paid
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateVacationPaidRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.CreateVacationPaidResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Create vacation paid
      tags:
      - VacationPaid
  /api/v1/vacations_paid/{vacation_paid_id}:
    delete:
      description: Delete vacation paid
      parameters:
      - description: vacation_paid_id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Delete vacation paid
      tags:
      - VacationPaid
  /api/v1/vacations_paid/{vacations_paid_id}:
    get:
      description: get vacation paid
      parameters:
      - description: vacation_paid_id
        in: path
        name: vacation_paid_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetVacationPaidResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Get vacation paid
      tags:
      - VacationPaid
  /api/v1/vacations_unpaid:
    post:
      description: Create vacation unpaid
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateVacationUnpaidRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.CreateVacationUnpaidResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Create vacation unpaid
      tags:
      - VacationUnpaid
  /api/v1/vacations_unpaid/{vacation_unpaid_id}:
    delete:
      description: Delete vacation unpaid
      parameters:
      - description: vacation_unpaid_id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Delete vacation unpaid
      tags:
      - VacationUnpaid
  /api/v1/vacations_unpaid/{vacations_unpaid_id}:
    get:
      description: get vacation unpaid
      parameters:
      - description: vacations_unpaid_id
        in: path
        name: vacation_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetVacationUnpaidResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperror.ErrorResponse'
      summary: Get vacation unpaid
      tags:
      - VacationUnpaid
swagger: "2.0"